Non-decreasing arrays_DSA_Hackerearth
Hey there!, Thanks for submitting a PR to this repo
Please complete the following checklist for your PR to be approved. Mark the items by adding an x.
Example: - [ ] Mark me as completed
To mark: - [x] Mark me as completed

My Pull Request adheres to the Contribution guidelines.
I have added readme.md and solution files in the appropriate folder.
The solution I have given is my own and not plagiarized in any form.
My solution passes all the testcases.

    I have properly indented and documented my solution using comments.

If any of the above is unmarked, please make the necessary changes and then mark them so that you're PR can be approved and merged.